2/27 2차 작업
Sssibal! 됐다!
플랫폼이 용암과 충돌했을 때 사라지지 않던 버그 해결.
플랫폼에 끼는 버그 있음.
플랫폼 랜덤 위치 소환 시스템을 좀 손봐야 함.

+++ Assets/Script/aboutPlatform.cs
@@ -62,12 +62,12 @@ public class aboutPlatform : MonoBehaviour
         }
     }
 
-    private int PlatformUp()
+    private float PlatformUp()
     {
         int yPosSpace = Random.Range(1,2);
 
         timer += Time.deltaTime;
         float PlatformSpace = timer * 1.5f + yPosSpace;
-        return (int) PlatformSpace;
+        return (float) PlatformSpace;
     }
 }
